fixing controller calls
