+++ webapi/csharp-ef-webapi/Controllers/MatchController.cs
@@ -1,8 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using DataAccessLibrary.Data;
-using DataAccessLibrary.Models.WebApi;
 using DataAccessLibrary.Models.GameCoordinator;
-using DataAccessLibrary.Models.ProMetadata;
 
 namespace csharp_ef_webapi.Controllers
 {
@@ -10,40 +8,20 @@ namespace csharp_ef_webapi.Controllers
     [ApiController]
     public class MatchController : ControllerBase
     {
-        private readonly GameCoordinatorRepository _gameCoordinatorRepository;
-        private readonly IMatchDetailRepository _matchDetailRepository;
-        private readonly ProMetadataRepository _proMetadataRepository;
+        private readonly IGcMatchMetadataRepository _gcMatchMetadataRepository;
 
         public MatchController(
-            GameCoordinatorRepository gameCoordinatorRepository,
-            IMatchDetailRepository matchDetailRepository,
-            ProMetadataRepository proMetadataRepository
+            IGcMatchMetadataRepository gcMatchMetadataRepository
         )
         {
-            _gameCoordinatorRepository = gameCoordinatorRepository;
-            _matchDetailRepository = matchDetailRepository;
-            _proMetadataRepository = proMetadataRepository;
-        }
-
-        // GET: api/Match/5/players
-        [HttpGet("{leagueId}/players")]
-        public async Task<ActionResult<IEnumerable<MatchDetailsPlayer>>> GetMatchPlayers(int leagueId)
-        {
-            League? league = await _proMetadataRepository.GetLeagueAsync(leagueId);
-
-            if (league == null)
-            {
-                return NotFound("League ID not found");
-            }
-
-            return Ok(await _matchDetailRepository.GetByLeagueAsync(league));
+            _gcMatchMetadataRepository = gcMatchMetadataRepository;
         }
 
         // GET: api/Match/5/Metadata
         [HttpGet("{matchId}/metadata")]
         public async Task<ActionResult<GcMatchMetadata>> GetLeagueMatchIdMetadata(long matchId)
         {
-            var match = await _gameCoordinatorRepository.GetMatchMetadataAsync(matchId);
+            var match = await _gcMatchMetadataRepository.GetByIdAsync(matchId);
 
             if (match == null)
             {
