+++ webapi/data-access-library/Models/DotaGameCoordinator/GcMatchMetadata.cs
@@ -2,6 +2,7 @@ namespace DataAccessLibrary.Models.GameCoordinator;
 
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using DataAccessLibrary.Models.ProMetadata;
 
 // EF mapping for CDOTAMatchMetadata
 [Table("dota_gc_match_metadata")]
@@ -10,6 +11,9 @@ public class GcMatchMetadata
     [Key]
     [Column("id")]
     public long Id { get; set; }
+
+    [System.Text.Json.Serialization.JsonIgnore]
+    public League League { get; set; } = null!;
 
     [Column("match_id")]
     public long MatchId { get; set; }
