+++ webapi/data-access-library/Models/DotaWebApi/MatchDetailsPicksBans.cs
@@ -10,11 +10,11 @@ public class MatchDetailsPicksBans
 {
     [Key]
     [Column("id")]
-    [JsonIgnore]
+    [System.Text.Json.Serialization.JsonIgnore]
     public int Id { get; set; }
 
-    [JsonIgnore]
-    public required MatchDetail Match { get; set; }
+    [System.Text.Json.Serialization.JsonIgnore]
+    public MatchDetail Match { get; set; } = null!;
 
     [Column("is_pick")]
     [JsonProperty("is_pick")]
