+++ webapi/data-access-library/Models/Fantasy/FantasyPlayer.cs
@@ -2,8 +2,8 @@ namespace DataAccessLibrary.Models.Fantasy;
 
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 using DataAccessLibrary.Models.ProMetadata;
-using Newtonsoft.Json;
 
 [Table("dota_fantasy_players")]
 public class FantasyPlayer
@@ -12,7 +12,9 @@ public class FantasyPlayer
     [Column("id")]
     [JsonIgnore]
     public long Id { get; set; }
-    public required FantasyLeague FantasyLeague { get; set; }
+
+    [JsonIgnore]
+    public FantasyLeague FantasyLeague { get; set; } = null!;
     public required Team Team { get; set; }
     public required Account DotaAccount { get; set; }
